// frontend/src/types/generated/Constraint.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Filter } from "./Filter";
import type { SizeFilter } from "./SizeFilter";

export type Constraint = { "type": "Filter", filter: Filter, } | { "type": "SizeFilter", filter: SizeFilter, } | { "type": "SAT", child_names: Array<string>, } | { "type": "ANY", child_names: Array<string>, } | { "type": "NOT", child_names: Array<string>, } | { "type": "OR", child_names: Array<string>, } | { "type": "AND", child_names: Array<string>, };